requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
gunicorn
python-multipart==0.0.9
pandas==2.2.1
google-cloud-documentai==2.25.0